Minor additions: Added hotkeys to GUI buttons, camera bookmarking (not sure if this system will stay), shoring up some JS error checking, moved a bunch of constants into the config file and added a new rotation mode.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@804 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/sysdep/win/wsdl.h
@@ -195,7 +195,7 @@ SDL_ActiveEvent;
 
 typedef struct
 {
-	u16 code;
+	intptr_t code;
 }
 SDL_UserEvent;
